# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Cinéphoria API
  version: 1.0.0
  description: Documentation de l’API REST Cinéphoria (Node.js, Express, PostgreSQL)
  contact:
    name: Redondo
    email: contact@cinephoria.fr
servers:
  - url: http://localhost:3000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /employee/films:
    get:
      summary: Récupère tous les films
      tags:
        - Employés
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste de tous les films
    post:
      summary: Crée un nouveau film
      tags:
        - Employés
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                titre:
                  type: string
                  example: Inception
                description:
                  type: string
                  example: Un film de science-fiction réalisé par Christopher Nolan.
                duree:
                  type: integer
                  example: 148
      responses:
        "201":
          description: Film créé avec succès
  "/employee/films/{id}":
    get:
      summary: Récupère un film par son ID
      tags:
        - Employés
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID du film
      responses:
        "200":
          description: Film trouvé
    put:
      summary: Met à jour un film
      tags:
        - Employés
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Film mis à jour
    delete:
      summary: Supprime un film
      tags:
        - Employés
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Film supprimé
  /employee/salles:
    get:
      summary: Récupère toutes les salles
      tags:
        - Employés
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste des salles
    post:
      summary: Crée une nouvelle salle
      tags:
        - Employés
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Salle créée
  "/employee/salles/{id}":
    get:
      summary: Récupère une salle par son ID
      tags:
        - Employés
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Salle trouvée
    put:
      summary: Met à jour une salle
      tags:
        - Employés
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Salle mise à jour
    delete:
      summary: Supprime une salle
      tags:
        - Employés
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Salle supprimée
  /employee/seances:
    get:
      summary: Récupère toutes les séances
      tags:
        - Employés
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste des séances
    post:
      summary: Crée une nouvelle séance
      tags:
        - Employés
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Séance créée
  "/employee/seances/{id}":
    get:
      summary: Récupère une séance par son ID
      tags:
        - Employés
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Séance trouvée
    put:
      summary: Met à jour une séance
      tags:
        - Employés
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Séance mise à jour
    delete:
      summary: Supprime une séance
      tags:
        - Employés
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Séance supprimée
  /employee/avis:
    get:
      summary: Récupère tous les avis
      tags:
        - Employés
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste de tous les avis
  /employee/avis/en-attente:
    get:
      summary: Récupère les avis en attente de validation
      tags:
        - Employés
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste des avis en attente
  "/employee/avis/{id}/valider":
    patch:
      summary: Valide un avis
      tags:
        - Employés
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Avis validé
  "/employee/avis/{id}":
    delete:
      summary: Supprime un avis
      tags:
        - Employés
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Avis supprimé
  /employee/dashboard:
    get:
      summary: Récupère les statistiques du tableau de bord employé
      tags:
        - Employés
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Statistiques du tableau de bord renvoyées
...